# documentation/api.yaml
openapi: 3.0.0
info:
  title: Westco Tack API Course Project Documentation.
  version: 1.0
  licence:
    name: Apache 2.0
    url: "http://apache.org/"
servers:
  - url: http://localhost:5000
    description: Dev Server on Port
  # - url: http://staging-api.example.com
  #   description: Optional server description, e.g. Internal staging server for testing
components:
  securitySchemes:
    bearerAuth: 
      type: http
      scheme: bearer
      bearerFormat: JWT 
security:
  - bearerAuth: []
paths:
...
  /getCourseOne/{id}:
    get:
      summary: Get One User
      tags:
        - Course
      description: Get one Course
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: Get Course One for course Id
          example: 62aec4b4edf4613e87f85cb6
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                tyep: object
                properties:
                _id: string
                name: string
                authorId: array
                price: number
                courseImg: string
                courseUsersId: array
                reating: number
                description: string
                lessons: array
                books: array
                comments: array
                createdAt: string
                updatedAt: string
              example:
                _id: 62ae948249ef0d92e405b22b
                name: Node.js darslari 0 dan pro gacha
                authorId: []
                price: 1500000
                courseImg: 18d181a6-d74b-40b3-b886-c409d76a6d21-file-nodejs_corse.png
                courseUsersId: []
                reating: 0
                description: ""
                lessons: []
                books: []
                comments: []
                createdAt: "2022-06-19T03:14:10.029Z"
                updatedAt: "2022-06-19T05:55:21.434Z"
        404:
          description: Not Found Course
        500:
          description: Internal Server Error
  /getCourseComment/{id}:
    get:
      summary: Get One Course Comment
      tags:
        - Course
      description: Get one Course Comment
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: Get Course One for course Id
          example: 62ae9ca688144d2d246e879d
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                tyep: object
                properties:
                _id: string
                name: string
                courseId: string
                userId: string
                description: string
                createdAt: string
                updatedAt: string
              example:
                "_id": "62ae9ca688144d2d246e879d"
                "courseId": "62ae95af49ef0d92e405b231"
                "name": "Sunna"
                "description": "Assalamu Aleykum menga bu React Next JS darslari juda maqul bo'ldi "
                "userId": "62ac7f3fa525215eeede092f"
                "createdAt": "2022-06-19T03:48:54.490Z"
                "updatedAt": "2022-06-19T03:48:54.490Z"
        404:
          description: Not Found Course
        500:
          description: Internal Server Error
  /getCourseBook/{id}:
    get:
      summary: Get One Course Comment
      tags:
        - Course
      description: Get one Course Comment
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: Get Course One for course Id
          example: 62ae9ca688144d2d246e879d
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                tyep: object
                properties:
                _id: string
                name: string
                courseId: string
                books: string
                urlBooks: string
                description: string
                createdAt: string
                updatedAt: string
              example:
                _id: string
                name: string
                courseId: string
                books: string
                urlBooks: string
                description: string
                createdAt: string
                updatedAt: string
        404:
          description: Not Found Course
        500:
          description: Internal Server Error
  /signUp:
    post:
      summary: Creates a Users
      tags:
        - Register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                password:
                  type: string
              example:
                name: Sunnatullox
                email: Sunnatullox@gmail.com
                password: 12345
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                tyep: object
                properties:
                _id: integer
                name: string
                email: string
                password: string
                userLessons: array
                userAdmin: boolen
                userCourseId: array
              example:
                _id: 62ae948249ef0d92e405b22b
                name: Sunnatullox
                email: Sunnatullox@gmail.com
                password: 12345
                userLessons: []
                userAdmin: false
                userCourseId: []
        404:
          description: Not Found Course
        500:
          description: Internal Server Error
  /signIn:
    post:
      summary: Sign In Users
      tags:
        - Register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
              example:
                email: Sunnatullox@gmail.com
                password: 12345
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                tyep: object
                properties:
                token: integer
                _id: integer
                name: string
                email: string
                password: string
                userLessons: array
                userAdmin: boolen
                userCourseId: array
              example:
                token: TU3ODExNjB9.RoCiHr5wflJbYUWs67NIK4IL3PQcM6fTDg2eeBGsJWo
                _id: 62ae948249ef0d92e405b22b
                name: Sunnatullox
                email: Sunnatullox@gmail.com
                password: $2a$10$4JUm3qBt.FPGxnqcfYZ08./loDWvYA8VpHxDYdMqVBccieZv/vuBua
                userLessons: []
                userAdmin: false
                userCourseId: []
        404:
          description: Not Found Course
        500:
          description: Internal Server Error
...
  /admin/courseCreate:
    post:
      summary: admin Create Course
      tags:
        - Admin Course
      description: Creates Course
      consumes:
        - multipart/form-data
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                authorId:
                  type: string
                name:
                  type: string
                price:
                  type: string
                description:
                  type: string
                courseImg:
                  type: string
                  format: binary
      responses:
        200:
          description: OK
          content:
            multipart/form-data:
              schema:
                tyep: object
                properties:
                adminProfile: string
              example:
                adminProfile: Secsess fully updated User Profile
        404:
          description: Not Found Course
        500:
          description: Internal Server Error
...
  /admin/courseDelete/{id}:
    delete:
      put:
      summary: Admin course delete
      tags:
        - Admin Course
      description: admin Delete course
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: this is the course ID to Admin deleted course
          example: 62b20de847fc3dfb699ec2a7
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                tyep: object
                properties:
                deletedCourse: string
              example:
                deletedCourse: "Admin sucsess fuly deleted cours"
        404:
          description: Not Found Course
        500:
          description: Internal Server Error
  /user/courseComment/create:
    post:
      summary: User Course Comment create
      tags:
        - User Course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                courseId:
                  type: string
                userId:
                  type: string
                name:
                  type: string
                description:
                  type: string
              example:
                courseId: 62aec4b4edf4613e87f85cb6
                userId: 62aeb9b09d6840a18ff77f47
                name: Sunnatullox
                description: Kurs Darslari qachon qo'yiladi
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                tyep: object
                properties:
                _id: integer
                courseId: integer
                name: string
                description: string
                userId: string
                createdAt: string
                updatedAt: string
              example:
                "_id": "62b216049920204eb56c309b"
                "userId": "62aeb9b09d6840a18ff77f47"
                "courseId": "62aec4b4edf4613e87f85cb6"
                "name": "Sunnatullox"
                "description": "Kurs Darslari qachon qo'yiladi?"
                "createdAt": "2022-06-21T19:03:32.308Z"
                "updatedAt": "2022-06-21T19:03:32.308Z"
        404:
          description: Not Found Course
        500:
          description: Internal Server Error
  /user/courseComment/update/{id}:
    put:
      put:
      summary: User course comment update
      tags:
        - User Course
      description: User Course Comment upate
      consumes:
        - multipart/form-data
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: this is the Comment ID to updated comment user
          example: 62b216049920204eb56c309b
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
              example:
                name: Sunnatullox
                description: Kurs Darslari qachon qo'yiladi
      responses:
        200:
          description: OK
          content:
            multipart/form-data:
              schema:
                _id: integer
                courseId: integer
                name: string
                description: string
                userId: string
                createdAt: string
                updatedAt: string
              example:
                "_id": "62b216049920204eb56c309b"
                "userId": "62aeb9b09d6840a18ff77f47"
                "courseId": "62aec4b4edf4613e87f85cb6"
                "name": "Sunnatullox"
                "description": "Kurs Darslari qachon qo'yiladi?"
                "createdAt": "2022-06-21T19:03:32.308Z"
                "updatedAt": "2022-06-21T19:03:32.308Z"
        404:
          description: Not Found Course
        500:
          description: Internal Server Error
  /user/courseComment/delete/{id}:
    delete:
      put:
      summary: User delete course comment
      tags:
        - User Course
      description: user Delete course commet
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: this is the course ID to Admin deleted course
          example: 62b216049920204eb56c309b
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                tyep: object
                properties:
                deletedCourse: string
              example:
                deletedCourse: "User sucsess fuly deleted cours comment"
        404:
          description: Not Found Course
        500:
          description: Internal Server Error
...